# improve rational performance